+++ src/bin/sks.ts
@@ -1,6 +1,6 @@
 #!/usr/bin/env node
 
-const FAST_PACKAGE_VERSION = '3.0.1';
+const FAST_PACKAGE_VERSION = '3.0.2';
 const args = process.argv.slice(2);
 
 try {